# src/presentation/http/session_tag_handler.rs
//! SessionTag REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{SessionTagService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateSessionTagDto, UpdateSessionTagDto, PatchSessionTagDto, SessionTagResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum SessionTagError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for SessionTagError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for SessionTagError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "SESSIONTAG_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "SESSIONTAG_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "SESSIONTAG_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "SESSIONTAG_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for SessionTag.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /session_tags | List with pagination |
/// | POST | /session_tags | Create new |
/// | GET | /session_tags/:id | Get by ID |
/// | PUT | /session_tags/:id | Full update |
/// | PATCH | /session_tags/:id | Partial update |
/// | DELETE | /session_tags/:id | Soft delete |
/// | POST | /session_tags/bulk | Bulk create |
/// | POST | /session_tags/upsert | Upsert |
/// | GET | /session_tags/trash | List deleted |
/// | POST | /session_tags/:id/restore | Restore |
/// | DELETE | /session_tags/empty | Empty trash |
/// | GET | /session_tags/:id/deleted | Get deleted by ID |
/// | DELETE | /session_tags/trash/:id | Permanent delete from trash |
/// | GET | /session_tags/count | Count active entities |
/// | GET | /session_tags/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(SessionTagService::with_repository(repository));
/// let router = create_session_tag_routes(service);
/// ```
pub fn create_session_tag_routes(service: Arc<SessionTagService>) -> Router {
    BackboneCrudHandler::<SessionTagService, SessionTag, CreateSessionTagDto, UpdateSessionTagDto, SessionTagResponseDto>::routes(
        service,
        "/session_tags",
    )
}

/// Create Axum router with only the read (GET) endpoints for SessionTag.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_session_tag_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_session_tag_read_routes(service: Arc<SessionTagService>) -> Router {
    BackboneCrudHandler::<SessionTagService, SessionTag, CreateSessionTagDto, UpdateSessionTagDto, SessionTagResponseDto>::read_routes(
        service,
        "/session_tags",
    )
}

/// Create Axum router with only the write (mutation) endpoints for SessionTag.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_session_tag_write_routes(service: Arc<SessionTagService>) -> Router {
    BackboneCrudHandler::<SessionTagService, SessionTag, CreateSessionTagDto, UpdateSessionTagDto, SessionTagResponseDto>::write_routes(
        service,
        "/session_tags",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_session_tag_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<SessionTagService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_session_tag_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}